backend/python-server/requirements.txt
# AllFileChanger – Unified Python Backend Dependencies
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# Image / PDF
img2pdf==0.5.1
Pillow>=10.0.0

# PDF manipulation
PyPDF2>=3.0.0
pdf2docx>=0.5.8

# Word / DOCX
python-docx>=1.1.0
docx2pdf>=0.1.8
docxcompose>=1.4.0

# Excel
pandas>=2.0.0
openpyxl>=3.1.0
tabula-py>=2.9.0

# PowerPoint
python-pptx>=0.6.23
pdf2image>=1.17.0

# OCR
pytesseract>=0.3.10

# System monitoring (MEMORY MANAGEMENT)
psutil>=5.9.0
